add Creator to Video, Change profile, password

+++ controller/videoController.js
@@ -1,7 +1,6 @@
 import routes from "../routes";
 import videoModel from "../models/Video";
 export const home = async (req, res) => {
-  console.log("hello");
   try {
     const videos = await videoModel.find({}).sort({ _id: -1 });
     res.render("home", { pageTitle: "home", videos });
@@ -9,7 +8,6 @@ export const home = async (req, res) => {
     console.log(error);
     res.render("home", { pageTitle: "home", videos: [] });
   }
-  console.log("hello");
 };
 
 export const search = async (req, res) => {
@@ -35,7 +33,14 @@ export const postUpload = async (req, res) => {
     file: { path: url },
   } = req;
   try {
-    const newVideo = await videoModel.create({ url, title, description });
+    const newVideo = await videoModel.create({
+      url,
+      title,
+      description,
+      creator: req.user.id,
+    });
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.video_detail(newVideo._id));
   } catch (error) {
     res.redirect(routes.home);
@@ -46,8 +51,7 @@ export const video_detail = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await videoModel.findById(id);
-    console.log(video);
+    const video = await videoModel.findById(id).populate("creator");
     res.render("video_detail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -57,10 +61,15 @@ export const edit_video = async (req, res) => {
   const {
     params: { id },
   } = req;
+  // router를 보호하기 위해서, 비디오의 creator와 loggedUser(접속된 유저로 쿠키에 저장된 값을 통해 식별)의 id가 다르면 Error를 발생시킴
   try {
     const video = await videoModel.findById(id);
-    console.log(video);
-    res.render("edit_video", { pageTitle: `Edit ${video.title}`, video });
+    if (String(video.creator) !== req.user.id) {
+      console.log("?");
+      throw Error();
+    } else {
+      res.render("edit_video", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -83,7 +92,12 @@ export const delete_video = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await videoModel.findOneAndRemove({ _id: id });
+    const video = await videoModel.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await videoModel.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
